requirements.txt
google-generativeai
python-dotenv
streamlit
langchain
PyPDF2
chromadb